requirements.txt
openai>=1.40.0
pydantic>=2.0.0
asyncache
cachetools
python-dotenv
fastapi
uvicorn[standard]
cachetools
slowapi
cryptography
httpx
python-multipart
supabase>=2.0.0